Header fixes, add weapon system adjusts into prey code

+++ app/src/main/jni/d3es-multithread-master/neo/game/Prey/game_slots.h
@@ -34,7 +34,7 @@
 	}fruit_t;
 
 	typedef struct victory_s {
-		int			f1, f2, f3;
+		long		f1, f2, f3;
 		int			payoff;
 	}victory_t;
